+++ Options.lua
@@ -11470,7 +11470,7 @@ local function CreateOptionsTable()
 
   -- Add dual-spec support, no also available in Classic (Season, Anniversary)
   local LibDualSpec = LibStub:GetLibrary("LibDualSpec-1.0", true)
-  if Addon.ExpansionIsAtLeastWrath or Addon.IS_CLASSIC_SOD or (C_Seasons and C_Seasons.GetActiveSeason() == 11) or (C_Seasons and C_Seasons.GetActiveSeason() == 12) then
+  if Addon.ExpansionIsAtLeastTBC or Addon.IS_CLASSIC_SOD or (C_Seasons and C_Seasons.GetActiveSeason() == 11) or (C_Seasons and C_Seasons.GetActiveSeason() == 12) then
     if LibDualSpec then
       LibDualSpec:EnhanceDatabase(Addon.db, t.ADDON_NAME)
       LibDualSpec:EnhanceOptions(options.args.profiles, Addon.db)