+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # @project-version@ (@build-time@)
 
+* Added support for WoW Burning Crusade Classic Anniversary Edition.
 * Fixed a Lua error that occurred in rare circumstances in WoW Classic when a unit's castbar was shown.
 * Support for dual spec profiles has been enabled for WoW Classic.
 * Fixed a bug in Auras widget that prevented buffs of hostile units from being displayed. [PM].